Validator checks data follow simple rules.
The data validator is meant to make the simpler part of checking whether
certain data conforms to the LTI specs. Right now, this is pretty much
just two things: whether a field exists and whether it matches a set of
restricted values.

It can also output warnings if a recommended field is missing and
restrict fields to certain primitive types, but I haven't decided if
those are actually good features to have yet.