Basic in memory storage for OAuth data.
OAuth, at the very least, needs a place to store the secret key and
value used for signature verification. This basic storage class
OAuthStore is meant as a base class that other classes can inherit from
in order to provide a more complex storage backend. The basic OAuthStore
just stores everything in memory as a dict.

Since the basic storage is just as a dict, the tests aren't very
complex.

Will probably need to add nonce storage to it too soon.