Implement OAuth signature verification.
Currently only validates the signature, does not enforce timestamp/nonce
policy.

The parser constructor parameter list is a bit long, due to OAuth
verification requiring so many pieces of the http request. Technically,
get_params and post_params can be merged into one piece, since OAuth
combines them all anyways. Not sure if this is the best way to do it
yet.

There's a lot little details of the spec that we have to support, I've
tried to heavily comment them when necessary. The tests should test all
of them that I'm consciously aware of.